docs: REST DOCS 운동 상세 루틴 가져오기 작성
