+++ src/docs/asciidoc/renewToken.adoc
@@ -1,4 +1,5 @@
 = 액세스 토큰 재발급
-+
+
+ㅇ
 
 operation::renew-access-token-test/renew-access-token-test[snippets='http-request,request-fields,http-response,response-fields']
