+++ src/docs/asciidoc/routine.adoc
@@ -1,4 +1,7 @@
 = 루틴
 
 === 카테고리별 루틴 목록 가져오기
-operation::routine-by-category/get-routine-by-category[snippets='http-request,path-parameters,response-body,response-fields']
+operation::routine-by-category-test/get-routine-by-category[snippets='http-request,path-parameters,response-body,response-fields']
+
+=== 상세 루틴 목록 가져오기
+operation::routine-detail-test/get-routine-detail[snippets='http-request,request-headers,path-parameters,response-body,response-fields']
