+++ src/main/resources/static/renewToken.html
@@ -5,7 +5,6 @@
 <meta http-equiv="X-UA-Compatible" content="IE=edge">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <meta name="generator" content="Asciidoctor 2.0.18">
-<meta name="author" content="+">
 <title>액세스 토큰 재발급</title>
 <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Open+Sans:300,300italic,400,400italic,600,600italic%7CNoto+Serif:400,400italic,700,700italic%7CDroid+Sans+Mono:400,700">
 <style>
@@ -441,11 +440,13 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="header">
 <h1>액세스 토큰 재발급</h1>
 <div class="details">
-<span id="author" class="author">+</span><br>
 <span id="revnumber">version 0.0.1-SNAPSHOT</span>
 </div>
 </div>
 <div id="content">
+<div class="paragraph">
+<p>ㅇ</p>
+</div>
 <div class="sect1">
 <h2 id="_http_request">HTTP request</h2>
 <div class="sectionbody">
@@ -457,7 +458,7 @@ Host: 54.180.88.182
 Content-Length: 195
 
 {
-  "refreshToken" : "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhb3NzdXBlcjdAbmF2ZXIuY29tIiwiZXhwIjoxNjk3ODc3MTQ3fQ.eDlxqzwXQT4TPk7V4oKt1sct5em3BlW0T6Rc3whycQ9WDEdegSjVF89fgU7VNMcikohN-RzGt_vdLgSSP9-kpQ"
+  "refreshToken" : "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhb3NzdXBlcjdAbmF2ZXIuY29tIiwiZXhwIjoxNjk3OTY2Mzk1fQ.L2Jd_9QD0lGVriBN4xNIFA8ZM9SgUw2z0xts6OMwOaPV7vIqjlZHUF_p97KSryOi8H0OWiNTcb-c0p0AzCbqPA"
 }</code></pre>
 </div>
 </div>
@@ -502,7 +503,7 @@ Content-Length: 303
   "code" : "200",
   "message" : "요청에 성공하였습니다.",
   "result" : {
-    "accessToken" : "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhb3NzdXBlcjdAbmF2ZXIuY29tIiwiZXhwIjoxNjk2NjY5MzQ3fQ.GxjJS3f3jxFSjNotJXyuN5m7pmVZTRA9_afD81vvAgE4KJ3bY5I21xscVXkF9sp8Gb4QDlx3Fs30d9X-wgkPTA"
+    "accessToken" : "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhb3NzdXBlcjdAbmF2ZXIuY29tIiwiZXhwIjoxNjk2NzYwNjMwfQ.hGCDEG9hQiBwsygY3phlBPc2LKLwZtXhT096llI2YbS3lTAyfHanEKHMkTFRPAdstenaZ4c5nRr5LHRByTtfKA"
   },
   "success" : true
 }</code></pre>
@@ -555,7 +556,7 @@ Content-Length: 303
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-10-05 16:46:56 +0900
+Last updated 2023-10-08 18:33:58 +0900
 </div>
 </div>
 </body>
