+++ src/test/java/project1/OurFit/signinUpTest/SignupTest.java
@@ -1,4 +1,4 @@
-package project1.OurFit;
+package project1.OurFit.signinUpTest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -18,26 +18,21 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.context.WebApplicationContext;
-import project1.OurFit.entity.RefreshToken;
-import project1.OurFit.jwt.JwtTokenProvider;
-import project1.OurFit.repository.RefreshTokenRedisRepository;
-import project1.OurFit.response.JwtTokenDto;
+import project1.OurFit.request.MemberDTO;
 import project1.RestDocsConfiguration;
 
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 
 @SpringBootTest
 @Transactional
 @ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
 @Import(RestDocsConfiguration.class)
-public class RenewAccessTokenTest {
+public class SignupTest {
     @Autowired private RestDocumentationResultHandler restDocs;
     private MockMvc mockMvc;
-    @Autowired private JwtTokenProvider jwtTokenProvider;
-    @Autowired private RefreshTokenRedisRepository refreshTokenRedisRepository;
     private ObjectMapper objectMapper = new ObjectMapper();
 
     @BeforeEach
@@ -53,28 +48,45 @@ public class RenewAccessTokenTest {
     }
 
     @Test
-    void renewAccessTokenTest() throws Exception {
-        //Given
-        RefreshToken refreshToken = refreshTokenRedisRepository.findByEmail("aossuper7@naver.com").get();
-        JwtTokenDto jwtTokenDto = new JwtTokenDto();
-        jwtTokenDto.setRefreshToken(refreshToken.getRefreshToken());
+    void signUp() throws Exception {
+        // Given
+        MemberDTO member = new MemberDTO();
+        member.setEmail("aossuper8@naver.com");
+        member.setPassword("aossuper8");
+        member.setNickname("aossuper8");
+        member.setGender("male");
+        member.setHeight(170.0);
+        member.setWeight(70.0);
+        member.setSquat(120.0);
+        member.setBenchpress(65.0);
+        member.setDeadlift(115.0);
+        member.setOverheadpress(40.0);
 
         //When & Then
-        mockMvc.perform(post("/newtoken")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectMapper.writeValueAsString(jwtTokenDto)))
+        mockMvc.perform(post("/signup")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(member)))
                 .andExpect(status().isOk())
                 .andDo(print())
                 .andDo(restDocs.document(
                         requestFields(
-                                fieldWithPath("refreshToken").description("리프레시 토큰")
+                                fieldWithPath("email").description("이메일"),
+                                fieldWithPath("password").description("패스워드"),
+                                fieldWithPath("nickname").description("닉네임"),
+                                fieldWithPath("gender").description("성별"),
+                                fieldWithPath("height").description("키"),
+                                fieldWithPath("weight").description("몸무게"),
+                                fieldWithPath("squat").description("스쿼트, 필수값 X").optional(),
+                                fieldWithPath("benchpress").description("벤치프레스, 필수값 X").optional(),
+                                fieldWithPath("deadlift").description("데드리프트, 필수값 X").optional(),
+                                fieldWithPath("overheadpress").description("오버헤드프레스, 필수값 X").optional()
                         ),
                         responseFields(
-                                fieldWithPath("code").description("Http 상태코드"),
-                                fieldWithPath("message").description("상태 메시지"),
-                                fieldWithPath("success").description("성공 여부"),
-                                fieldWithPath("result.accessToken").description("재발급 받은 액세스 토큰")
+                                fieldWithPath("success").description("성공여부"),
+                                fieldWithPath("code").description("HTTP 상태 코드"),
+                                fieldWithPath("message").description("메시지")
                         )
                 ));
     }
+
 }
